update bg-white, add sweetalert.

+++ src/components/Customers/Customer.svelte
@@ -1,31 +1,51 @@
 <script>
+  import Swal from 'sweetalert2';
   export let customerData;
 
-  function handleDeleteCustumer(customerId) {
-    console.log(customerId);
+  function handleDeleteCustumer(customerId, customerName) {
+    console.log(customerId, customerName);
 
-// firebase.firestore()
-    //   .collection('customers')
-    //   .doc(customerId)
-    //   .get()
-    //   .then((snapshot) => console.log(snapshot.size));
+    // firebase.firestore()
+        //   .collection('customers')
+        //   .doc(customerId)
+        //   .get()
+        //   .then((snapshot) => console.log(snapshot.size));
 
-    // let query = firebase.firestore().collection('customers');
+        // let query = firebase.firestore().collection('customers');
 
-    // query.get().then(querySnapshot => {
-    //   if (querySnapshot.docs.length) {
-    //     console.log('No documents found.', querySnapshot.docs.length);
-    //   }
-    // });
+        // query.get().then(querySnapshot => {
+        //   if (querySnapshot.docs.length) {
+        //     console.log('No documents found.', querySnapshot.docs.length);
+        //   }
+        // });
+  Swal.fire({
+    title: `${customerName}:
+            Are you sure? `,
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        console.log(result.value)
+        firebase.firestore().collection("customers")
+          .doc(customerId)
+          .delete().then(() => {
+            console.log("Document successfully deleted:", customerId);
+            Swal.fire(
+              'Deleted!',
+              'Your Customer has been deleted.',
+              'success'
+            );
+          }).catch((error) => {
+            console.error("Error removing document: ", error);
+          });
+      }
+  })
 
-  firebase.firestore().collection("customers")
-    .doc(customerId)
-    .delete().then(() => {
-      console.log("Document successfully deleted:", customerId);
-    }).catch((error) => {
-      console.error("Error removing document: ", error);
-    });
-  }
+}
 
 </script>
 
@@ -42,6 +62,9 @@
       </div>
       <div class="flex flex-col text-right md:text-right">
         <a rel='prefetch' href='customers/{customerData.id}'><i class="fas fa-sun text-orange-600 pr-3"></i></a>
+        <button on:click={handleDeleteCustumer(customerData.id, customerData.data().name)}>
+          <i class="far fa-trash-alt text-orange-600 pr-3"></i>
+        </button>
       </div>
     </div>
   </div>
